GSM irrigation system
This is a project in which I am remotely controlling the irrigation system via SMSs through a cellular GSM network. I have used the mcb(Miniature Circuit Breaker) here which is use to detect any electric fault or motor jamming and also Send back ERROR messages to the operator.the code for the same is provided here.