Add files via upload
This is quite simple project which is use to detect the obstacle with the help of the ultrasonic sensor and stop the vehicle. For this project i have use arduino uno, arduino motor driver shield and the ultrasonic sensor.The ultrasonic sensor is use to detect the obstacle in-front of it and provide the appropriate distance between the car and the obstacle.